Add string primitive operations, though actual foreign types need to be defined independently at the user level.
To accomodate some of the demands of a string primitive, make it
possible to pass an arbitrary number of arguments to a structure mutator
and accessor. This allows you to pass explicit size arguments when
necessary. I don't know the affect on speed this will have, but I don't
think it should be too bad.